+++ src/DotNetWorker.ApplicationInsights/release_notes.md
@@ -1,4 +1,5 @@
 ## What's Changed
 
-### Microsoft.Azure.Functions.Worker.ApplicationInsights 1.2.0
-- Updating `Microsoft.ApplicationInsights.PerfCounterCollector` to 2.22.0
+### Microsoft.Azure.Functions.Worker.ApplicationInsights 1.3.0
+- Updating `Azure.Identity` to 1.11.0
+- Enabling support for Microsoft Entra authentication for Application Insights